# Speed up Go implementation