Unroll inner loop and postpone shifting.

```
benchmark                                 old ns/op      new ns/op     delta
BenchmarkCount8/generic/1000-32           5049           2008          -60.23%
BenchmarkCount8/generic/10000-32          50975          20084         -60.60%
BenchmarkCount8/generic/100000-32         510629         201260        -60.59%
BenchmarkCount8/generic/1000000-32        5186954        2013423       -61.18%
BenchmarkCount8/generic/10000000-32       51857148       20017886      -61.40%
BenchmarkCount8/generic/100000000-32      515500200      200000200     -61.20%
BenchmarkCount16/generic/1000-32          14440          3793          -73.73%
BenchmarkCount16/generic/10000-32         145768         38286         -73.73%
BenchmarkCount16/generic/100000-32        1451220        381340        -73.72%
BenchmarkCount16/generic/1000000-32       15645678       3799999       -75.71%
BenchmarkCount16/generic/10000000-32      154999986      38032245      -75.46%
BenchmarkCount16/generic/100000000-32     1544998200     379333367     -75.45%

benchmark                                 old MB/s     new MB/s     speedup
BenchmarkCount8/generic/1000-32           198.06       497.94       2.51x
BenchmarkCount8/generic/10000-32          196.17       497.92       2.54x
BenchmarkCount8/generic/100000-32         195.84       496.87       2.54x
BenchmarkCount8/generic/1000000-32        192.79       496.67       2.58x
BenchmarkCount8/generic/10000000-32       192.84       499.55       2.59x
BenchmarkCount8/generic/100000000-32      193.99       500.00       2.58x
BenchmarkCount16/generic/1000-32          138.51       527.22       3.81x
BenchmarkCount16/generic/10000-32         137.20       522.38       3.81x
BenchmarkCount16/generic/100000-32        137.82       524.47       3.81x
BenchmarkCount16/generic/1000000-32       127.83       526.32       4.12x
BenchmarkCount16/generic/10000000-32      129.03       525.87       4.08x
BenchmarkCount16/generic/100000000-32     129.45       527.24       4.07x
```