@fuzxxl 2020-08-28
Thank you for your contributions  Do you agree to contribute these under the terms and conditions of the 2-clause BSD license as used in this project (I'll add the license in a moment)?

And would you mind adding implementations for Count32 and Count64 as well?

I'll then change the testing code to test everything against the “safe” functions instead of the “generic” functions.

One thing I noted is that you use `uint64` for the temporary counter vector.  This may lead to suboptimal performance on 32 bit targets where `int` is a 32 bit type.  Is there any particular reason for this choice?