@klauspost 2020-08-28
> And would you mind adding implementations for Count32 and Count64 as well?

Done.

> I'll then change the testing code to test everything against the “safe” functions instead of the “generic” functions.

Did that to test.

> One thing I noted is that you use uint64 for the temporary counter vector. This may lead to suboptimal performance on 32 bit targets where int is a 32 bit type. Is there any particular reason for this choice?

Mostly to have control over the length. On 32 bits you would limit yourself to 15 or 16 bits anyway, so it would be a tradeoff. But it can be tweaked to work better on 32 bits and you could of course split up the input.

TBH 32 bit is getting so rare that I don't really target it. Amazed you 'bothered' to add 386 support :)