@klauspost 2020-08-28
FWIW this is the speed difference on `386`:
```
benchmark                                 old ns/op      new ns/op     delta
BenchmarkCount8/generic/1000-32           6413           3894          -39.28%
BenchmarkCount8/generic/10000-32          64882          38510         -40.65%
BenchmarkCount8/generic/100000-32         646210         387840        -39.98%
BenchmarkCount8/generic/1000000-32        6521979        3854105       -40.91%
BenchmarkCount8/generic/10000000-32       65944450       38766653      -41.21%
BenchmarkCount8/generic/100000000-32      660999850      385333233     -41.70%
BenchmarkCount16/generic/1000-32          13145          7716          -41.30%
BenchmarkCount16/generic/10000-32         132611         77294         -41.71%
BenchmarkCount16/generic/100000-32        1322686        773924        -41.49%
BenchmarkCount16/generic/1000000-32       13235311       7727270       -41.62%
BenchmarkCount16/generic/10000000-32      132250038      77214257      -41.61%
BenchmarkCount16/generic/100000000-32     1322999800     767000150     -42.03%

benchmark                                 old MB/s     new MB/s     speedup
BenchmarkCount8/generic/1000-32           155.93       256.81       1.65x
BenchmarkCount8/generic/10000-32          154.13       259.68       1.68x
BenchmarkCount8/generic/100000-32         154.75       257.84       1.67x
BenchmarkCount8/generic/1000000-32        153.33       259.46       1.69x
BenchmarkCount8/generic/10000000-32       151.64       257.95       1.70x
BenchmarkCount8/generic/100000000-32      151.29       259.52       1.72x
BenchmarkCount16/generic/1000-32          152.15       259.22       1.70x
BenchmarkCount16/generic/10000-32         150.82       258.75       1.72x
BenchmarkCount16/generic/100000-32        151.21       258.42       1.71x
BenchmarkCount16/generic/1000000-32       151.11       258.82       1.71x
BenchmarkCount16/generic/10000000-32      151.23       259.02       1.71x
BenchmarkCount16/generic/100000000-32     151.17       260.76       1.72x
```