@fuzxxl 2020-08-28
I believe the same or even better performance is achievable on 32 bit targets as on 64 bit targets if 32 bit counters are used.  One solution for all architectures would be to use an inner loop with `uint32` counters and an outer one with `int` counters.  The inner one iterates as much as possible until overflow can no longer be avoided, then the outer loop accumulates the partial results into the `int` sized counters and continues until the array has been exhausted.