@klauspost 2020-08-28
> Do you agree to contribute these under the terms and conditions of the 2-clause BSD license as used in this project (I'll add the license in a moment)?

Yes.