@fuzxxl 2020-08-28
Rebased and merged manually.